requirements.txt
absl-py==1.3.0
altair==4.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-core==4.11.1
jupyter_client==7.4.2
kaggle==1.5.12
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
pydot==1.4.2
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
stack-data==0.5.1
streamlit==1.15.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
